Configure Particle System

+++ BattleTanks 4.21 - 2/Source/BattleTanks/Public/TankAimingComponent.h
@@ -11,7 +11,7 @@
 
 // Enum for Aiming state
 UENUM()
-enum class EFiringStatus : uint32
+enum class EFiringStatus : uint8
 {
 	Locked,
 	Aiming,
